Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic<2.0.0

metaapi-cloud-sdk
aiohttp
python-dotenv
